+++ src/main/java/com/project/service/abstraction/BookService.java
@@ -42,5 +42,5 @@ public interface BookService {
 
     List<BookNewDTO>getAllLightBookDtoForSearch();
 
-    int getSizeOfTotalBooks();
+    Long getSizeOfTotalBooks();
 }
